apps/requirements.txt
pandas
numpy
matplotlib
plotly
fastparquet
ipywidgets
dash
pyyaml
